Added Autograd NP
Also added Hybrid version.

Cleaned up code.
