+++ Non parametric/Frechet/Tutorial - available kernels
