+++ Non parametric/Frechet/Tutorial - how to use
